(aoe): rename package name
