+++ pkg/vm/engine/aoe/storage/db/sched/flushseg.go
@@ -1,4 +1,4 @@
-package db
+package sched
 
 import (
 	"matrixone/pkg/container/batch"
