+++ pkg/vm/engine/aoe/storage/db/sched/metamgr.go
@@ -1,4 +1,4 @@
-package db
+package sched
 
 import (
 	"errors"
