+++ pkg/vm/engine/aoe/storage/db/sched/sched_test.go
@@ -1,4 +1,4 @@
-package db
+package sched
 
 import (
 	"github.com/stretchr/testify/assert"
