+++ pkg/vm/engine/aoe/storage/db/sched/scheduler.go
@@ -1,4 +1,4 @@
-package db
+package sched
 
 import (
 	e "matrixone/pkg/vm/engine/aoe/storage"
