+++ pkg/vm/engine/aoe/storage/db/sched/types_test.go
@@ -1,4 +1,4 @@
-package db
+package sched
 
 import (
 	e "matrixone/pkg/vm/engine/aoe/storage"
